tests/notebook_requirements.txt
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.28.0
openai==1.12.0
python-dotenv==1.0.0
asyncpg==0.29.0
pandas==2.2.0
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0  # For potential experimentation with embeddings and similarity
networkx==3.2.1  # For potential graph-based relationship analysis 